[form-factors] Correct expressions of the factors of the radiative correction factor
The factor C needs to be evaluated at energy scale mu_h1 and the factor K needs to be evaluated at energy scale mu_h2. The factor U2 needs to be corrected replacing the anomalous dimension of the SCET heavy-light current gamma_i by the anomalous dimension of the heavy-light current in HQET gamma_i_hl.